Function AddStudentDetails connected to Database

+++ src/com/company/User.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 public class User {
     private int studentID;
     private String studentName;
-    private int studentGPA;
+    private Double studentGPA;
     private ArrayList<Subject> subjects = new ArrayList<>();
 
 
@@ -19,11 +19,11 @@ public class User {
         this.studentID = studentID;
     }
 
-    public int getStudentGPA() {
+    public Double getStudentGPA() {
         return studentGPA;
     }
 
-    public void setStudentGPA(int studentGPA) {
+    public void setStudentGPA(Double studentGPA) {
         this.studentGPA = studentGPA;
     }
 
@@ -43,7 +43,7 @@ public class User {
         return subjects;
     }
 
-    public User(int studentID , String studentName, int studentGPA){
+    public User(int studentID , String studentName, double studentGPA){
         this.studentID = studentID;
         this.studentName = studentName;
         this.studentGPA = studentGPA;
